minesweeper/requirements.txt
emoji==2.6.0
Pillow==10.0.0